Accept null value for the sort option
To not introduce a breaking change in the API of the translation:extract
command, I need to accept the "null" value for the `--sort` option when
`--force` is passed to the command. Meaning that I had to remove the
default value of the `InputOption`.

In the mean time, I didn't want to break the previous behaviour when
using `--dump-messages`. So in this case, when the option is not passed,
I fallback to the value "asc".