[Serializer] Rewrite `AbstractObjectNormalizer::createChildContext()` to use the provided cache_key from original context when creating child contexts
If the normalization starts with an initial 'cache_key' it is currently discarded when creating child contexts. This change fixes that.
The main reason behind it is that if the client code sends a unique identifier (ApiPlatform injects the IRI) and we have a use case
that iterates a big resultset we end up with a big private cache that quickly runs out of allowed memory. The solution should be to
send a 'cache_key' which ignores the entire cache key calculation that hashes the context.