[HttpKernel] Add `#[MapSessionParameter]` to pass a session parameter to a controller argument
This attribute allow to pass an object from the session.
The read/write to the public properties are proxied to the session using the property name as a key, or the value defined by the attribute SessionKey

This allows to type property and have ide autocompletion, static validation and runtime validation of type.
The default value of the argument is also used as a default value when not defined in session.

In case of interfaces, the user will be required to provide a default value or make the parameter nullable.
This check is done even if the session may have a value already to notify quickly of the potential issue