[DoctrineBridge] Add argument to EntityValueResolver to set type aliases
This allows for fixing https://github.com/symfony/symfony/issues/51765; with a consequential Doctrine bundle update, the resolve_target_entities configuration can be injected similarly to ResolveTargetEntityListener in the Doctrine codebase.

Alternatively the config and ValueResolver can be injected using a compiler pass in the Symfony core code, however the value resolver seems to be configured in the Doctrine bundle already.