[WebProfilerBundle] Add settings to have the profiler follow the browser.
Often during development, we load a page, check some informations in the profiler, see something wrong, fix, reload the page, reload the profiler.
That mean that we often reload the profiler page.
If we can use the 'latest' token it's ok, but if there is another controller called (for example a webmanifest), we either need to search for it or close the tab, then reopen a new tab.

This change use the BroadcastChannel javascript api to broadcast to other tabs that a toolbar have loaded or that an ajax request have finished.
The profiler page listen to this broadcast and show a toolbar with the new profiler page based on the token broadcasted.
A new setting allow to choose if :
- We don't want this behavior
- We want to follow only the main requests where a toolbar is shown (default)
- We also want to follow ajax call captured by the toolbar