[Cache] Stop defaulting to igbinary in `DefaultMarshaller`
igbinary used to be a drop in replacement for PHP's `serialize` but recent changes (in PHP 7.4) to the handling of uninitialized properties in `serialize` have not made it into igbinary, so it no longer is a simple drop in replacement.

This only removes igbinary as the default serializer, code can still opt-in through the first constructor argument.

This may result in a performance regression on systems with igbinary as it would no longer be used by default.