terminate with non-zero exit code when a secret could not be read
The fix for #42038 did a bit too much. Not only did it fix the PHP error that
wasn't caught before, but also changed the exit code of the command. With this
change the PHP error will still be prevented, but the command will terminate
with a non-zero exit code to indicate the failure that occurred while reading
the stored secrets.