fix handling empty data in ValueToDuplicatesTransformer
The transformer receives the data of child forms that have already been
through the transformation schema. If no custom view transformer was used
on that child form the empty would already have been changed to null.
Thus, receiving an empty string here means that the child form explicitly
asked for it and the value must not be exchanged with null.