Add conflict with doctrine/data-fixtures >= 2
The signature of ContainerAwareLoader::addFixture() is not compatible with the
v2 signature of the Loader interface from doctrine/data-fixtures, as per this
fatal error:

> Declaration of Symfony\Bridge\Doctrine\DataFixtures\ContainerAwareLoader::addFixture(Doctrine\Common\DataFixtures\FixtureInterface $fixture)
>              must be compatible with Doctrine\Common\DataFixtures\Loader::addFixture(Doctrine\Common\DataFixtures\FixtureInterface $fixture): void

The class is not final, which means adding the return type declaration would be
a breaking change.

Let us add a conflict to signal this temporary incompatibility.