Named serializer should get correct normalizer instance
Named serializers were introduced in #56823 and they work great.

We noticed a small bug when using custom name convertors.

The MimeMessageNormalizer holds a reference to `serializer.normalizer.property`.
But when using named serializers, it would use the specific child normalizer instead.

With this change, we fix this problem for any service that starts with `serializer.normalizer.`.