Create an `AsArgumentValueResolver` to allow setting priority
When creating a value resolver, there is a decent change you need to set the priority.
Especially if you need to resolve a request attribute, as the priority needs to be
higher than the `RequestAttributeValueResolver`.

Currently, the only way to do this is by setting the priority in the `services.yaml`.
This PR introduces the `AsArgumentValueResolver` attribute, which allows setting the
`name` and `priority` directly on the class with an attribute, rather than in the config.