[Messenger] Allow handler locator to be set by applications
The `HandleMessageMiddleware` has a dependency on a `HandlersLocatorInterface`
but this gets registered by the messenger pass.

As far as I can see there isn't a way for an application to provide it's
own implementation. By checking if the service already exists before
registering then we can allow applications to define their own