Implement batch size limits
Some transports will have a max batch limit (eg. SQS)
Either the transport has to make sub batches, or the messageBus should fail when trying to send bigger batches
The issue with the transport making sub batches, is that you have the chance that half the batch fails and half succeeds.
eg.
dispatchBatch(20 messages)
- transport sends 10
- transport tries to send 10 again, but fails

now half the batch got sent, but we have no clue which ones.
IMO, best to block this.